CPi.Lib

Documentation

data CpiException

Constructors

CpiException String

data Species

Constructors

Nil

Def String [Name]

Sum [PrefixSpecies]

Par [Species]

New AffNet Species

data Process

Constructors

Process [(Species, Conc)] AffNet

data Prefix

Constructors

Comm Name OutNames InNames

Tau Rate

type PrefixSpecies = (Prefix, Species)

data AffNet

Constructors

AffNet [Aff]

data Aff

Constructors

Aff ((Name, Name), Rate)

data Definition

Constructors

SpeciesDef Name [Name] Species

ProcessDef Name Process

type Conc = Double

type Env = [Definition]

type Name = String

type OutNames = [Name]

type InNames = [Name]

type Rate = Double

class Show a => Pretty a where

Methods

pretty :: a -> String

class Nf a where

Methods

nf :: a -> a

prettys :: Pretty a => [a] -> String

Pretty print a list of pretty printable expressions.

prettyNames :: [Name] -> String

Pretty print a list of Names.

prettyList :: [[Char]] -> [Char]

Pretty print a list

aff :: AffNet -> (Name, Name) -> Maybe Rate

tri1 :: (t, t1, t2) -> t

First element of a triple.

tri2 :: (t, t1, t2) -> t1

Second element of a triple.

tri3 :: (t, t1, t2) -> t2

Third element of a triple.

s2d :: String -> Double

d2s :: Double -> String

lookupDef :: Env -> Species -> Maybe Species

revLookupDef :: Env -> Species -> Maybe Species

lookupProcName :: Env -> String -> Maybe Process

lookupSpecName :: Env -> String -> Maybe Species

speciesInProc :: Process -> [Species]

compose :: Process -> Process -> Process

card :: Eq a => a -> [a] -> Integer

Count of an element in a list.

sites :: AffNet -> [Name]

sub :: [(Name, Name)] -> Species -> Species

remove :: Eq a => a -> [a] -> [a]

Remove first matched element of a list.

replace :: Eq a => a -> a -> [a] -> [a]

Replace first matched element of a list with something else.

ifnotnil

Arguments

:: [a]
List

-> ([a] -> b)
Function

-> b
Default

-> b

If List is not nil then apply Function else return Default.

netUnion :: AffNet -> AffNet -> AffNet

fn :: Species -> [Name]

bn :: Species -> [Name]

rename :: (Name, Name) -> Species -> Species

vecRename :: (Name, Name) -> [Name] -> [Name]

netRename :: (Name, Name) -> AffNet -> AffNet

specName :: Species -> String

infty :: Double

(/\) :: Eq a => [a] -> [a] -> [a]

Set intersection.

(\/) :: Eq a => [a] -> [a] -> [a]

Set union.

(#) :: AffNet -> Species -> Bool

(##) :: AffNet -> AffNet -> Bool

(\\) :: Eq a => [a] -> [a] -> [a]

Set difference.