CPi.Semantics

Important functions:

processMTS :: Env -> Process -> MTS

Get the Multi-Transition System for a Process.

Datatypes:

data Concretion

Constructors

ConcBase Species OutNames InNames

ConcPar Concretion [Species]

ConcNew AffNet Concretion

data MTS

Constructors

MTS [Trans]

data Trans

Constructors

TransSC Species Name Concretion

TransT Species TTau Species

TransTA Species TTauAff Species

data TTau

Constructors

TTau Rate

data TTauAff

Constructors

TTauAff (Name, Name)

Functions:

lookupTrans :: MTS -> Species -> [Trans]

Lookup the transitions (in an MTS) from a species.

pseudoapp :: Concretion -> Concretion -> Maybe Species

Pseudo-application of concretions. Takes takes two concretions and if they're compatable then gives the species they combine to form.

potentials :: MTS -> [Trans]

Gives the Class 1 (Species-->Concretion) transitions of an MTS.

initconc :: Process -> Species -> Double

Initial concentration of a species in a process.

primes :: Env -> Species -> [Species]

Prime components of a species, i.e. species that are not compositions.

wholeProc :: Env -> Process -> MTS -> Process

Gives the complete syntactic process, including all potentially generated primes.